Sometimes, the file name is too long for yt-dlp to download
When the file name includes an emoji, the length is calculated incorrectly, and yt-dlp needs extra space for the temporary file name. This is fixed by reducing the file name length to 140 bytes.

Since the file name length has changed, in order to avoid downloading all files again, the script now checks if a file is already downloaded by examining only the first 50 characters.